Issue #0300: fixed release module

+++ DSLAMAll/DSLAMBackend/src/main/java/com/imotion/dslam/backend/persistence/jpa/DSLAMBKPersistenceServiceBaseJPA.java
@@ -6,6 +6,7 @@ import com.imotion.dslam.backend.persistence.DSLAMBKPersistenceServiceBase;
 import com.imotion.dslam.backend.persistence.service.file.DSLAMBKFilePersistenceServiceJPA;
 import com.selene.arch.exe.back.persistence.AEMFTIPersistenceService;
 import com.selene.arch.exe.back.persistence.module.AEMFTIPersistenceModule;
+import com.selene.arch.exe.core.AEMFTICoreProxyService;
 
 public abstract class DSLAMBKPersistenceServiceBaseJPA<T, Q extends T, Id extends Serializable> extends DSLAMBKPersistenceServiceBase<T, Q, Id> {
 
@@ -17,29 +18,40 @@ public abstract class DSLAMBKPersistenceServiceBaseJPA<T, Q extends T, Id extend
 	
 	
 	@Override
-	public AEMFTIPersistenceModule<Q, Id> getPersistenceModule() {
+	public DSLAMBKPersistenceModuleJPA<Q, Id> getPersistenceModule() {
 		if (persistenceModule == null) {
 			persistenceModule = new DSLAMBKPersistenceModuleJPA<Q, Id>();
 			persistenceModule.initialize(new Object[] { getPersistenceCoreService(), getPersistenceClass()});
 		}
-		persistenceModule.createEntityManager();
 		return persistenceModule;
 	}
 	
 	@Override
 	public void releaseModule() {
-		persistenceModule = null;
+		if (persistenceModule != null) {
+			persistenceModule.destroyCurrentEntityManager();
+			persistenceModule.releaseInstance();
+			persistenceModule = null;
+		}
 	}
+	
+	/**************************************************************
+     *                AEMFTIFACTORABLE FUNCTIONS                  *
+     **************************************************************/
 	
 	/**************************************************************
      *                AEMFTIFACTORABLE FUNCTIONS                  *
      **************************************************************/
 
 	@Override
+	public void initialize(Object[] args) {
+		super.initialize(args);
+		getPersistenceModule().createEntityManager();
+	}
+
+	@Override
 	public void releaseInstance() {
 		super.releaseInstance();
-		persistenceModule.destroyCurrentEntityManager();
-		persistenceModule 		= null;
 		if (filePersistence != null) {
 			getFactoryPersistence().release((AEMFTIPersistenceService<?, ?, ?>) filePersistence);
 			filePersistence = null;
